+++ Quick.ZLMediaKit/WebHook/Model/StreamNotFoundInfo.cs
@@ -25,7 +25,7 @@ namespace Quick.ZLMediaKit.WebHook.Model
         /// <summary>
         /// 播放器端口号
         /// </summary>
-        public ushort Port { get; set; }
+        public int Port { get; set; }
         /// <summary>
         /// 播放的协议，可能是rtsp、rtmp
         /// </summary>
